requirements.txt
anyio==3.6.2
azure-common==1.1.28
azure-core==1.26.2
azure-functions==1.12.0
azure-mgmt-core==1.3.2
azure-mgmt-rdbms==10.1.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
Cython==0.29.33
fastapi==0.88.0
gevent==22.10.2
google-api-core==2.11.0
google-api-python-client==2.74.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.58.0
greenlet==2.0.2
grequests==0.6.0
h11==0.14.0
httplib2==0.21.0
idna==3.4
injector==0.20.1
isodate==0.6.1
logging-azure==1.0.2
msrest==0.7.1
numpy==1.22.4
oauth2client==4.1.3
oauthlib==3.2.2
pandas==1.4.2
protobuf==4.21.12
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.3
PyDrive==1.3.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
subprocess32==3.5.4
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.14
uvicorn==0.20.0
zope.event==4.6
zope.interface==5.5.2
python-dotenv